# openapi.yaml
openapi: 3.0.3
info:
  title: BankWiz API
  version: 0.1.0
  description: An API for accessing banking functionalities provided by BankWiz
servers:
  - url: http://localhost:8080
    description: Local development (server)
...
security:
  - oauth2:
    - openid
    - profile
    - email
paths:
...
  /bankaccount:
    post:
      tags:
        - BankAccountService
      summary: Create a bank account
      operationId: createBankAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankAccountCreationRequest'
      responses:
        '201':
          description: Bank account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountIndexDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /bankaccount/{id}:
    delete:
      tags:
        - BankAccountService
      summary: Delete a bank account
      operationId: deleteBankAccount
      parameters:
        - name: id
          in: path
          description: Bank account ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bank account deleted
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
    put:
      tags:
        - BankAccountService
      summary: Update a bank account
      operationId: updateBankAccount
      parameters:
        - name: id
          in: path
          description: Bank account ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankAccountUpdateRequest'
      responses:
        '200':
          description: Bank account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountIndexDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
  /bankaccount/bankaccounts:
    get:
      tags:
        - BankAccountService
      summary: Get all bank account
      operationId: getAllBankAccount
      responses:
        '200':
          description: Bank account created successfully
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/GroupBankAccountIndexDTO'
        '400':
          description: Invalid request. Please check the provided data.
        default:
          description: Unexpected error
#################################### BankAccountService END ######################### 
#################################### TransactionService START ####################### 
...
components:
  securitySchemes:
    oauth2:   # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: oauth2
      description: For more information, see https://api.slack.com/docs/oauth
      flows: 
        implicit:
          authorizationUrl: https://bankwiz-dev.eu.auth0.com/authorize?audience=Bankwiz_server
          scopes:
            openid: openid
            profile: profile
            email: email
  schemas:
...
    BankAccountCreationRequest:
      type: object
      properties:
        bankAccountName:
          type: string
        groupId:
          type: string
          format: uuid
        decimalBaseAmount:
          type: integer
          format: int32
        currency:
          $ref: '#/components/schemas/CurrencyEnum'
      required:
        - bankAccountName
        - groupId
        - decimalBaseAmount
        - currency
    BankAccountUpdateRequest:
      type: object
      properties:
        bankAccountName:
          type: string
        groupId:
          type: string
          format: uuid
        decimalBaseAmount:
          type: integer
          format: int32
...
    GroupIndexDTO:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
      required:
        - groupId
        - groupName
...
    BankAccountIndexDTO:
      type: object
      properties:
        bankAccountName:
          type: string
        bankAccountId:
          type: string
          format: uuid
        decimalBaseAmount:
          type: integer
          format: int32
        currency:
          $ref: '#/components/schemas/CurrencyIndexDTO'
      required:
        - bankAccountName
        - bankAccountId
        - decimalBaseAmount
        - currency
    GroupBankAccountIndexDTO:
      type: object
      properties:
        groupeIndex:
          $ref: '#/components/schemas/GroupIndexDTO'
        bankAccountIndexList:
          type: "array"
          items:
            $ref: '#/components/schemas/BankAccountIndexDTO'
      required:
        - groupeIndex
        - bankAccountIndexList
...
    CurrencyIndexDTO:
      type: object
      properties:
        currencyIsoCode3:
          type: string
        displayName:
          type: string
        symbol:
          type: string
      required:
        - currencyIsoCode3
        - displayName
        - symbol
#################################### DTO END ######################################## 
#################################### Enum START #####################################
...
    CurrencyEnum:
      type: string
      enum:
        - EUR
        - USD
#################################### Enul END #######################################